requirements.txt
lolapy @ git+https://github.com/alejamp/lola-py-sdk 
python-dotenv==1.0.0
Flask==2.3.2